lesson04/devops10/requirements.txt
amqp==2.6.0
billiard==3.6.3.0
casbin==0.8.4
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-jsonfield==1.4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-extensions==0.6.0
future==0.18.2
greenlet==0.4.16
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
packaging==20.4
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.7
pytz==2020.1
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
shortuuid==1.0.1
simpleeval==0.9.10
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
